requirements.txt
soundfile
torchaudio
# kokoro @ git+https://huggingface.co/hexgrad/Kokoro-82M@main
fastapi
pydantic
uvicorn
pydub
python-magic
scipy
transformers
munch
phonemizer
